fix delete and undo actions
