+++ app/client/src/layoutSystems/anvil/integrations/actions/actionTypes.ts
@@ -34,7 +34,6 @@ export enum AnvilReduxActionTypes {
   ANVIL_MOVE_WIDGET = "ANVIL_MOVE_WIDGET",
   ANVIL_ADD_SUGGESTED_WIDGET = "ANVIL_ADD_SUGGESTED_WIDGET",
   ANVIL_SECTION_ZONES_UPDATE = "ANVIL_SECTION_ZONES_UPDATE",
-  SAVE_ANVIL_LAYOUT = "SAVE_ANVIL_LAYOUT",
   ANVIL_SPACE_DISTRIBUTION_START = "ANVIL_SPACE_DISTRIBUTION_START",
   ANVIL_SPACE_DISTRIBUTION_UPDATE = "ANVIL_SPACE_DISTRIBUTION_UPDATE",
   ANVIL_SPACE_DISTRIBUTION_STOP = "ANVIL_SPACE_DISTRIBUTION_STOP",
